refactor: move test files to a new tests folder at root
