+++ __tests__/min-length-validation.spec.ts
@@ -1,6 +1,6 @@
 import faker from '@faker-js/faker';
 import { InvalidFieldError } from '~/validation/errors';
-import { MinLengthValidation } from './min-length-validation';
+import { MinLengthValidation } from '~/validation/validators/min-length/min-length-validation';
 
 const makeSut = (length: number) => {
   const field = faker.database.column();
