+++ src/main/factories/http/axios-http-client.factory.ts
@@ -1,3 +1,3 @@
-import { AxiosHttpClient } from '~/infra/http/axios-http-client/axios-http-client';
+import { AxiosHttpClient } from '~/infra/http/axios-http-client';
 
 export const makeAxiosHttpClient = () => new AxiosHttpClient();
